Add SABnzb
Added sabnzb -- Able to download via two different methods. If you run the it normally with one argument it will try to automatically detect current version, and check for a new version by adding 1 to the last octet of the version numbers. If a new version is detected its downloaded, else it exits the script. The second method is by passing a download URL as the second argument.